+++ ios/Classes/MobilepayAppswitchPlugin.h
@@ -1,7 +1,5 @@
 #import <Flutter/Flutter.h>
 #import <MobilePay-AppSwitch-SDK/MobilePayManager.h>
 
-@class MobilepayAppswitchPlugin;
-
 @interface MobilepayAppswitchPlugin : NSObject<FlutterPlugin>
 @end