+++ ios/Classes/MobilepayAppswitchPlugin.m
@@ -1,7 +1,8 @@
 #import "MobilepayAppswitchPlugin.h"
+#import <mobilepay_appswitch/mobilepay_appswitch-Swift.h>
 
 @implementation MobilepayAppswitchPlugin
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [MobilepayAppswitchPlugin registerWithRegistrar:registrar];
+    [SwiftMobilepayAppswitchPlugin registerWithRegistrar:registrar];
 }
 @end